+++ fjs/djs/parser/module.f.ts
@@ -9,7 +9,8 @@ import type { List } from '../../types/list/types.ts'
 import { fold, next, toArray, length, concat } from '../../types/list/module.f.mjs'
 import type { Fold } from '../../types/function/operator/types.ts'
 import type { DjsToken, DjsTokenWithMetadata } from '../tokenizer/module.f.ts'
-import { setReplace, at, type OrderedMap } from '../../types/ordered_map/module.f.ts'
+import { setReplace, at } from '../../types/ordered_map/module.f.mjs'
+import type { OrderedMap } from '../../types/ordered_map/types.ts'
 import { fromMap } from '../../types/object/module.f.ts'
 import type { AstArray, AstConst, AstModule, AstModuleRef } from '../ast/module.f.ts'
 import type { TokenMetadata } from '../../js/tokenizer/module.f.ts'
