+++ fjs/djs/transpiler/module.f.ts
@@ -9,7 +9,8 @@ import { error, ok } from '../../types/result/module.f.mjs'
 import type { List } from '../../types/list/types.ts'
 import { drop, map as listMap, toArray, includes } from '../../types/list/module.f.mjs'
 import { tokenize } from '../tokenizer/module.f.ts'
-import { setReplace, at, type OrderedMap } from '../../types/ordered_map/module.f.ts'
+import { setReplace, at } from '../../types/ordered_map/module.f.mjs'
+import type { OrderedMap } from '../../types/ordered_map/types.ts'
 import { stringToList } from '../../text/utf16/module.f.mjs'
 import { concat as pathConcat } from '../../path/module.f.ts'
 import { type ParseError, parseFromTokens } from '../parser/module.f.ts'
