+++ fjs/js/tokenizer/module.f.ts
@@ -11,7 +11,8 @@ import type { RangeMapArray, RangeMerge } from '../../types/range_map/types.ts'
 import { merge, fromRange, get } from '../../types/range_map/module.f.mjs'
 import type { List } from '../../types/list/types.ts'
 import { empty, stateScan, flat, toArray, reduce as listReduce, scan, map as listMap } from '../../types/list/module.f.mjs'
-import { at, fromEntries, type Entry } from '../../types/ordered_map/module.f.ts'
+import { at, fromEntries } from '../../types/ordered_map/module.f.mjs'
+import type { Entry } from '../../types/ordered_map/types.ts'
 import type { Range as NumberRange } from '../../types/range/types.ts'
 import { one } from '../../types/range/module.f.mjs'
 import type { BigFloat } from '../../types/bigfloat/types.ts'
