+++ fjs/types/ordered_map/types.ts
@@ -1,0 +1,5 @@
+import type { Tree } from "../btree/types/types.ts"
+
+export type Entry<T> = readonly [string, T]
+
+export type OrderedMap<T> = Tree<Entry<T>>
